+++ include/Gaffer/FileSequencePathFilter.h
@@ -77,7 +77,7 @@ class GAFFER_API FileSequencePathFilter : public PathFilter
 
 	protected :
 
-		void doFilter( std::vector<PathPtr> &paths ) const override;
+		void doFilter( std::vector<PathPtr> &paths, const IECore::Canceller *canceller ) const override;
 
 	private :