+++ include/Gaffer/LeafPathFilter.h
@@ -59,7 +59,7 @@ class GAFFER_API LeafPathFilter : public Gaffer::PathFilter
 
 	protected :
 
-		void doFilter( std::vector<PathPtr> &paths ) const override;
+		void doFilter( std::vector<PathPtr> &paths, const IECore::Canceller *canceller ) const override;
 
 	private :