+++ include/Gaffer/MatchPatternPathFilter.h
@@ -74,7 +74,7 @@ class GAFFER_API MatchPatternPathFilter : public Gaffer::PathFilter
 
 	protected :
 
-		void doFilter( std::vector<PathPtr> &paths ) const override;
+		void doFilter( std::vector<PathPtr> &paths, const IECore::Canceller *canceller ) const override;
 
 	private :