+++ src/Gaffer/FileSystemPath.cpp
@@ -89,7 +89,7 @@ FileSystemPath::~FileSystemPath()
 {
 }
 
-bool FileSystemPath::isValid() const
+bool FileSystemPath::isValid( const IECore::Canceller *canceller ) const
 {
 	if( !Path::isValid() )
 	{
@@ -105,7 +105,7 @@ bool FileSystemPath::isValid() const
 	return t != status_error && t != file_not_found;
 }
 
-bool FileSystemPath::isLeaf() const
+bool FileSystemPath::isLeaf( const IECore::Canceller *canceller ) const
 {
 	return isValid() && !is_directory( path( this->string() ) );
 }
@@ -146,11 +146,12 @@ FileSequencePtr FileSystemPath::fileSequence() const
 	}
 
 	FileSequencePtr sequence = nullptr;
+	/// \todo Add cancellation support to `ls`.
 	IECore::ls( this->string(), sequence, /* minSequenceSize = */ 1 );
 	return sequence;
 }
 
-void FileSystemPath::propertyNames( std::vector<IECore::InternedString> &names ) const
+void FileSystemPath::propertyNames( std::vector<IECore::InternedString> &names, const IECore::Canceller *canceller ) const
 {
 	Path::propertyNames( names );
 
@@ -165,15 +166,17 @@ void FileSystemPath::propertyNames( std::vector<IECore::InternedString> &names )
 	}
 }
 
-IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedString &name ) const
+IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedString &name, const IECore::Canceller *canceller ) const
 {
 	if( name == g_ownerPropertyName )
 	{
 		if( m_includeSequences )
 		{
+			IECore::Canceller::check( canceller );
 			FileSequencePtr sequence = fileSequence();
 			if( sequence )
 			{
+				IECore::Canceller::check( canceller );
 				std::vector<std::string> files;
 				sequence->fileNames( files );
 
@@ -182,6 +185,7 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 				std::map<std::string,size_t> ownerCounter;
 				for( std::vector<std::string>::iterator it = files.begin(); it != files.end(); ++it )
 				{
+					IECore::Canceller::check( canceller );
 					struct stat s;
 					stat( it->c_str(), &s );
 					struct passwd *pw = getpwuid( s.st_uid );
@@ -208,9 +212,11 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 	{
 		if( m_includeSequences )
 		{
+			IECore::Canceller::check( canceller );
 			FileSequencePtr sequence = fileSequence();
 			if( sequence )
 			{
+				IECore::Canceller::check( canceller );
 				std::vector<std::string> files;
 				sequence->fileNames( files );
 
@@ -219,6 +225,7 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 				std::map<std::string,size_t> ownerCounter;
 				for( std::vector<std::string>::iterator it = files.begin(); it != files.end(); ++it )
 				{
+					IECore::Canceller::check( canceller );
 					struct stat s;
 					stat( it->c_str(), &s );
 					struct group *gr = getgrgid( s.st_gid );
@@ -247,15 +254,18 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 
 		if( m_includeSequences )
 		{
+			IECore::Canceller::check( canceller );
 			FileSequencePtr sequence = fileSequence();
 			if( sequence )
 			{
+				IECore::Canceller::check( canceller );
 				std::vector<std::string> files;
 				sequence->fileNames( files );
 
 				std::time_t newest = 0;
 				for( std::vector<std::string>::iterator it = files.begin(); it != files.end(); ++it )
 				{
+					IECore::Canceller::check( canceller );
 					std::time_t t = last_write_time( path( *it ), e );
 					if( t > newest )
 					{
@@ -276,15 +286,18 @@ IECore::ConstRunTimeTypedPtr FileSystemPath::property( const IECore::InternedStr
 
 		if( m_includeSequences )
 		{
+			IECore::Canceller::check( canceller );
 			FileSequencePtr sequence = fileSequence();
 			if( sequence )
 			{
+				IECore::Canceller::check( canceller );
 				std::vector<std::string> files;
 				sequence->fileNames( files );
 
 				uintmax_t total = 0;
 				for( std::vector<std::string>::iterator it = files.begin(); it != files.end(); ++it )
 				{
+					IECore::Canceller::check( canceller );
 					uintmax_t s = file_size( path( *it ), e );
 					if( !e )
 					{
@@ -318,7 +331,7 @@ PathPtr FileSystemPath::copy() const
 	return new FileSystemPath( names(), root(), const_cast<PathFilter *>( getFilter() ), m_includeSequences );
 }
 
-void FileSystemPath::doChildren( std::vector<PathPtr> &children ) const
+void FileSystemPath::doChildren( std::vector<PathPtr> &children, const IECore::Canceller *canceller ) const
 {
 	path p( this->string() );
 
@@ -329,15 +342,18 @@ void FileSystemPath::doChildren( std::vector<PathPtr> &children ) const
 
 	for( directory_iterator it( p ), eIt; it != eIt; ++it )
 	{
+		IECore::Canceller::check( canceller );
 		children.push_back( new FileSystemPath( it->path().string(), const_cast<PathFilter *>( getFilter() ), m_includeSequences ) );
 	}
 
 	if( m_includeSequences )
 	{
+		IECore::Canceller::check( canceller );
 		std::vector<FileSequencePtr> sequences;
 		IECore::ls( p.string(), sequences, /* minSequenceSize */ 1 );
 		for( std::vector<FileSequencePtr>::iterator it = sequences.begin(); it != sequences.end(); ++it )
 		{
+			IECore::Canceller::check( canceller );
 			std::vector<FrameList::Frame> frames;
 			(*it)->getFrameList()->asList( frames );
 			if ( !is_directory( path( (*it)->fileNameForFrame( frames[0] ) ) ) )