+++ src/Gaffer/MatchPatternPathFilter.cpp
@@ -104,7 +104,7 @@ bool MatchPatternPathFilter::getInverted() const
 	return m_inverted;
 }
 
-void MatchPatternPathFilter::doFilter( std::vector<PathPtr> &paths ) const
+void MatchPatternPathFilter::doFilter( std::vector<PathPtr> &paths, const IECore::Canceller *canceller ) const
 {
 	paths.erase(
 		std::remove_if(