+++ src/GafferScene/ScenePath.cpp
@@ -145,18 +145,22 @@ const Gaffer::Context *ScenePath::getContext() const
 	return m_context.get();
 }
 
-bool ScenePath::isValid() const
+bool ScenePath::isValid( const IECore::Canceller *canceller ) const
 {
 	if( !Path::isValid() )
 	{
 		return false;
 	}
 
-	Context::Scope scopedContext( m_context.get() );
+	Context::EditableScope scopedContext( m_context.get() );
+	if( canceller )
+	{
+		scopedContext.setCanceller( canceller );
+	}
 	return m_scene->exists( names() );
 }
 
-bool ScenePath::isLeaf() const
+bool ScenePath::isLeaf( const IECore::Canceller *canceller ) const
 {
 	// Any part of the scene could get children at any time
 	return false;
@@ -167,9 +171,19 @@ PathPtr ScenePath::copy() const
 	return new ScenePath( m_scene, m_context, names(), root(), const_cast<PathFilter *>( getFilter() ) );
 }
 
-void ScenePath::doChildren( std::vector<PathPtr> &children ) const
+const Gaffer::Plug *ScenePath::cancellationSubject() const
 {
-	Context::Scope scopedContext( m_context.get() );
+	return m_scene.get();
+}
+
+void ScenePath::doChildren( std::vector<PathPtr> &children, const IECore::Canceller *canceller ) const
+{
+	Context::EditableScope scopedContext( m_context.get() );
+	if( canceller )
+	{
+		scopedContext.setCanceller( canceller );
+	}
+
 	ConstInternedStringVectorDataPtr childNamesData = m_scene->childNames( names() );
 	const std::vector<InternedString> &childNames = childNamesData->readable();
 	ScenePlug::ScenePath childPath( names() );